+++ bin/dartnet.dart
@@ -8,6 +8,7 @@ import 'package:dartnet/dartnet.dart';
 main(List<String> args) async {
   ArgParser parser = new ArgParser();
   parser.addCommand("init", new ArgParser()..addOption("filename", abbr: "f", defaultsTo: dartnetConfigurationFile));
+  parser.addCommand("dockerize", new ArgParser()..addOption("filename", abbr: "f", defaultsTo: dartnetConfigurationFile));
   parser.addOption("config", abbr: "c", defaultsTo: dartnetConfigurationFile);
   parser.addFlag("help", abbr: "h", defaultsTo: false, negatable: false);
   ArgResults results;
@@ -17,6 +18,8 @@ main(List<String> args) async {
 
 COMMANDS:
 \tinit\tCreate config file with default value.
+\t\t${parser.commands["init"].usage}
+\tdockerize\tCreate a Dockerfile from the Dartnet config file
 \t\t${parser.commands["init"].usage}
     ''';
 
@@ -30,6 +33,8 @@ COMMANDS:
     print(usage);
   } else if (results.command?.name == "init") {
     initConfigFile(filename: results.command["filename"]);
+  } else if (results.command?.name == "dockerize") {
+    dockerize(filename: results.command["filename"]);
   } else {
     await start(configPath: results["config"]);
   }